apps/mentormap/backend/requirements.txt
# FastAPI - Updated for Python 3.13 compatibility
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1

# Validation - Updated for Python 3.13
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Payment
stripe==11.2.0
